tsu-nav-project/tsu_upskill_backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
PyJWT==2.8.1
google-generativeai==0.3.0
django-ratelimit==4.1.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
gunicorn==21.2.0
whitenoise==6.6.0
cryptography==41.0.7
django-filter==23.5
